Upload for the practice of Camera in OpenGL.
Read and add some comment in the code.
Try to compile and get the result , make me know more about camera.
Simply , it is using global variable in OpenGL - "uniform" in shader ,
and change it when user move the cursor or press w,a,s,d key in the game
loop.